// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        HistoryResponse: {
            /** @description 失败摘要。 */
            error_message?: string | null;
            /**
             * Format: date-time
             * @description 结束时间。
             */
            finished_at?: string | null;
            /**
             * Format: int64
             * @description 主键。
             */
            id: number;
            /**
             * Format: int64
             * @description 所属流水线。
             */
            pipeline_id: number;
            /**
             * Format: date-time
             * @description 开始时间。
             */
            started_at: string;
            /** @description 运行状态。 */
            status: string;
            /** @description 触发来源。 */
            trigger_source: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}